Abstract out spreadsheet parsing for testing
