+++ bmh_lims/frontend/src/assets/utils/csvReader.js
@@ -1,5 +1,14 @@
 import XLSX from 'xlsx'
 
+export const csvToJSON = dataText => {
+    const lines = dataText.trim().split('\n').map(line => {
+        const replacedCommas = line.replace(/"([^,"]*,[^,"]*)+"/gm, quote => quote.replace(/,/gm, '%%%%'))
+        const replacedQuotes = replacedCommas.replace(/"[^"]*""[^"]*""[^"]*"/gm, quote => quote.replace(/""/gm, '&&&&').replace(/"/gm, ''))
+        const values = replacedQuotes.split(',')
+        return values.map(value => value.replace(/%%%%/gm, ',').replace(/&&&&/gm, '"'))
+    })
+    return lines
+}
 
 export const csvReader = (csvFile, toDo) => {
     const reader = new FileReader()
@@ -18,7 +27,7 @@ export const xlsxReader = (xlsxFile, toDo) => {
         const wb = XLSX.read(bstr, {type:'binary'});
         const wsname = wb.SheetNames[0];
         const ws = wb.Sheets[wsname];
-        const data = XLSX.utils.sheet_to_csv(ws, {header:1});
+        const data = XLSX.utils.sheet_to_json(ws, {header:1});
         toDo(data)
     };
     reader.readAsBinaryString(xlsxFile);
