+++ bmh_lims/frontend/src/assets/utils/index.js
@@ -1,1 +1,1 @@
-export { csvReader, xlsxReader } from './csvReader'
+export { csvReader, xlsxReader, csvToJSON } from './csvReader'
